// frontend2/src/api/orval/client/model/canvasTextObject.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Military Journal API
 * OpenAPI spec version: 1.0.0
 */
import type { CanvasObjectProperties } from './canvasObjectProperties';
import type { CanvasObjectStyle } from './canvasObjectStyle';
import type { CanvasObjectTransform } from './canvasObjectTransform';
import type { TextField } from './textField';

export interface CanvasTextObject {
  id: string;
  type: 'text';
  transform: CanvasObjectTransform;
  style?: CanvasObjectStyle | null;
  properties?: CanvasObjectProperties | null;
  fields: TextField;
}